Updated Forwarding's unit test to expect a priority of 1 now in all flowmods.

+++ src/main/java/net/floodlightcontroller/routing/IRoutingDecision.java
@@ -61,6 +61,6 @@ public interface IRoutingDecision {
     public void setMulticastInterfaces(List<SwitchPort> lspt);
     public Match getMatch();
     public void setMatch(Match match);
-    public short getHardTimeout();
+    public int getHardTimeout();
     public void setHardTimeout(short hardTimeout);
 }
